Add some more defines

+++ src/util/Vector3f.hpp
@@ -29,7 +29,7 @@ public:
     }
     Vector3f operator-(const Vector3f &v) const
     {
-        return Vector3f(v.x - (float)x, v.y - (float)y, (float)v.z - z);
+        return Vector3f(v.x - x, v.y - y, v.z - z);
         //return Vector3f(x - v.x, v.y - y, v.z - z);
     }
